+++ public/samples/exports.js
@@ -1,0 +1,5 @@
+define(function (require, exports, module) {
+    exports['foo'] = 'bar';
+    exports['require'] = require;
+    exports['module'] = module;
+});