+++ src/main.ts
@@ -1,1 +1,3 @@
-export * from 'lib/dummy.ts';
+export { AMDLoader } from 'lib/amd-loader.ts';
+
+export type { SystemJS } from 'lib/systemjs.d.ts';